+++ delivery-market/src/main/java/com/ddangme/dm/service/cash/CashService.java
@@ -5,8 +5,8 @@ import com.ddangme.dm.exception.DMException;
 import com.ddangme.dm.exception.ErrorCode;
 import com.ddangme.dm.model.cash.CashCharging;
 import com.ddangme.dm.model.member.Member;
-import com.ddangme.dm.repository.MemberRepository;
 import com.ddangme.dm.repository.cash.CashRepository;
+import com.ddangme.dm.repository.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
